+++ src/get/_get.js
@@ -1,6 +1,7 @@
 
 // exporting modules to be included the UMD bundle
 
+import activeElement from './active-element';
 import activeElements from './active-elements';
 import focusRedirectTarget from './focus-redirect-target';
 import focusTarget from './focus-target';
@@ -9,6 +10,7 @@ import parents from './parents';
 import shadowHostParents from './shadow-host-parents';
 import shadowHost from './shadow-host';
 export default {
+  activeElement,
   activeElements,
   focusRedirectTarget,
   focusTarget,